+++ src/modules/customer/component/QuantitySlot.tsx
