+++ src/routes/RouteGuard.tsx
@@ -1,6 +1,7 @@
 import { Navigate, Outlet, useLocation, useNavigate } from 'react-router-dom';
 import { useAuthStore } from '../modules/auth/store/authStore';
 import { useEffect } from 'react';
+import { UserRole } from '../common/enums/UserRole';
 
 
 const RouteGuard = () => {
@@ -13,9 +14,10 @@ const RouteGuard = () => {
     }
 
     if (location.pathname === '/') {
-        if (userDto.role[0] === 0) {
+        const userRole = userDto.role[0];
+        if (userRole == UserRole.SELLER) {
             return <Navigate to="/seller" replace />;
-        } else if (userDto.role[0] === 1) {
+        } else if (userRole == UserRole.CUSTOMER) {
             return <Navigate to="/c" replace />;
         }
